requirements.txt
numpy==1.18.1
pandas==1.0.1
scipy==1.4.1
matplotlib==3.1.3
orange3==3.23.1
scikit_posthocs==0.6.4
plotly
dlib
scikit-video
opencv-python==4.2.0.34